+++ src/main/java/uniqram/c1one/comment/dto/CommentCreateRequest.java
@@ -9,7 +9,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class CommentCreateRequest {
     private Long userId; //임시 포함
-    private Long postId;
     private Long parentCommentId; //대댓글인 경우
     private String content;
 }
